✨ feat(badge): add clickable badge, move into folder
